web_scraping_com_selenium/requirements.txt
# Dependências principais para automação de navegador
selenium
webdriver-manager

# Para gerenciamento de variáveis de ambiente (.env)
python-dotenv